Downgrade electron-store to 8.2
This prevents the issue described here: https://github.com/sindresorhus/electron-store/issues/276

The recommended solution that issue is to convert the entire project to
ESM, which I've attempted in #183, but it doesn't seem to work because
of incompatibilities between Webpack and Electron Forge that I haven't
been able to resolve. So for now I'm just going to downgrade.
(The app works fine with or without the errors.)